Fix bug 530842 - Upgrade Sunbird/Lightning 0.9 to 1.0pre fails [error: no such table: idx_cal_alarms_cal_id_item_id_recurrence_id_recurrence_id_tz] (regression). r=mschroeder,ssitter

+++ calendar/providers/storage/calStorageCalendar.js
@@ -214,9 +214,11 @@ calStorageCalendar.prototype = {
                         if (this.mDB.tableExists("cal_events")) { // check again (with lock)
                             // take over data and drop from storage.sdb tables:
                             for (let table in getSqlTable(DB_SCHEMA_VERSION)) {
-                                this.mDB.executeSimpleSQL("CREATE TABLE local_sqlite." +  table +
-                                                          " AS SELECT * FROM " + table +
-                                                          "; DROP TABLE IF EXISTS " +  table);
+                                if (table.substr(0, 4) != "idx_") {
+                                    this.mDB.executeSimpleSQL("CREATE TABLE local_sqlite." +  table +
+                                                              " AS SELECT * FROM " + table +
+                                                              "; DROP TABLE IF EXISTS " +  table);
+                                }
                             }
                             this.mDB.commitTransaction();
                         } else { // migration done in the meantime
